Add Mock API class to unit tests
